Move code into abstract parent class.
