+++ src/main/java/edu/harvard/seas/pl/formulog/eval/RoundBasedStratumEvaluator.java
@@ -30,7 +30,6 @@ import edu.harvard.seas.pl.formulog.Configuration;
 import edu.harvard.seas.pl.formulog.ast.BindingType;
 import edu.harvard.seas.pl.formulog.ast.Term;
 import edu.harvard.seas.pl.formulog.ast.UserPredicate;
-import edu.harvard.seas.pl.formulog.ast.Var;
 import edu.harvard.seas.pl.formulog.db.SortedIndexedFactDb;
 import edu.harvard.seas.pl.formulog.eval.SemiNaiveRule.DeltaSymbol;
 import edu.harvard.seas.pl.formulog.symbols.RelationSymbol;
@@ -47,26 +46,24 @@ import edu.harvard.seas.pl.formulog.validating.ast.SimplePredicate;
 
 public final class RoundBasedStratumEvaluator extends AbstractStratumEvaluator {
 
-	final int stratumNum;
-	final SortedIndexedFactDb db;
-	SortedIndexedFactDb deltaDb;
-	SortedIndexedFactDb nextDeltaDb;
-	final CountingFJP exec;
-	final Set<RelationSymbol> trackedRelations;
-	volatile boolean changed;
+	private final int stratumNum;
+	private final SortedIndexedFactDb db;
+	private SortedIndexedFactDb deltaDb;
+	private SortedIndexedFactDb nextDeltaDb;
+	private final Set<RelationSymbol> trackedRelations;
+	private volatile boolean changed;
 
-	static final int taskSize = Configuration.taskSize;
-	static final int smtTaskSize = Configuration.smtTaskSize;
+	private static final int taskSize = Configuration.taskSize;
+	private static final int smtTaskSize = Configuration.smtTaskSize;
 
 	public RoundBasedStratumEvaluator(int stratumNum, SortedIndexedFactDb db, SortedIndexedFactDb deltaDb,
 			SortedIndexedFactDb nextDeltaDb, Iterable<IndexedRule> rules, CountingFJP exec,
 			Set<RelationSymbol> trackedRelations) {
-		super(rules);
+		super(rules, exec);
 		this.stratumNum = stratumNum;
 		this.db = db;
 		this.deltaDb = deltaDb;
 		this.nextDeltaDb = nextDeltaDb;
-		this.exec = exec;
 		this.trackedRelations = trackedRelations;
 	}
 
@@ -106,7 +103,8 @@ public final class RoundBasedStratumEvaluator extends AbstractStratumEvaluator {
 		}
 	}
 
-	void reportFact(RelationSymbol sym, Term[] args, Substitution s) throws EvaluationException {
+	@Override
+	protected void reportFact(RelationSymbol sym, Term[] args, Substitution s) throws EvaluationException {
 		Term[] newArgs = new Term[args.length];
 		for (int i = 0; i < args.length; ++i) {
 			newArgs[i] = args[i].normalize(s);
@@ -124,7 +122,7 @@ public final class RoundBasedStratumEvaluator extends AbstractStratumEvaluator {
 		}
 	}
 
-	void updateDbs() {
+	private void updateDbs() {
 		StopWatch watch = recordDbUpdateStart();
 		for (RelationSymbol sym : nextDeltaDb.getSymbols()) {
 			if (nextDeltaDb.isEmpty(sym)) {
@@ -142,7 +140,7 @@ public final class RoundBasedStratumEvaluator extends AbstractStratumEvaluator {
 	}
 
 	@SuppressWarnings("serial")
-	class UpdateDbTask extends AbstractFJPTask {
+	private class UpdateDbTask extends AbstractFJPTask {
 
 		final RelationSymbol sym;
 		final Iterator<Iterable<Term[]>> it;
@@ -164,7 +162,9 @@ public final class RoundBasedStratumEvaluator extends AbstractStratumEvaluator {
 
 	}
 
-	Iterable<Iterable<Term[]>> lookup(IndexedRule r, int pos, OverwriteSubstitution s) throws EvaluationException {
+	@Override
+	protected Iterable<Iterable<Term[]>> lookup(IndexedRule r, int pos, OverwriteSubstitution s)
+			throws EvaluationException {
 		SimplePredicate predicate = (SimplePredicate) r.getBody(pos);
 		int idx = r.getDbIndex(pos);
 		Term[] args = predicate.getArgs();
@@ -189,177 +189,10 @@ public final class RoundBasedStratumEvaluator extends AbstractStratumEvaluator {
 		return Util.splitIterable(ans, targetSize);
 	}
 
-	static final boolean recordRuleDiagnostics = Configuration.recordRuleDiagnostics;
-
 	@SuppressWarnings("serial")
-	class RuleSuffixEvaluator extends AbstractFJPTask {
+	private class RulePrefixEvaluator extends AbstractFJPTask {
 
-		final IndexedRule rule;
-		final SimplePredicate head;
-		final SimpleLiteral[] body;
-		final int startPos;
-		final OverwriteSubstitution s;
-		final Iterator<Iterable<Term[]>> it;
-
-		protected RuleSuffixEvaluator(IndexedRule rule, SimplePredicate head, SimpleLiteral[] body, int pos,
-				OverwriteSubstitution s, Iterator<Iterable<Term[]>> it) {
-			super(exec);
-			this.rule = rule;
-			this.head = head;
-			this.body = body;
-			this.startPos = pos;
-			this.s = s;
-			this.it = it;
-			if (Configuration.recordWork) {
-				Configuration.workItems.increment();
-			}
-		}
-
-		protected RuleSuffixEvaluator(IndexedRule rule, int pos, OverwriteSubstitution s,
-				Iterator<Iterable<Term[]>> it) {
-			super(exec);
-			this.rule = rule;
-			this.head = rule.getHead();
-			SimpleLiteral[] bd = new SimpleLiteral[rule.getBodySize()];
-			for (int i = 0; i < bd.length; ++i) {
-				bd[i] = rule.getBody(i);
-			}
-			this.body = bd;
-			this.startPos = pos;
-			this.s = s;
-			this.it = it;
-			if (Configuration.recordWork) {
-				Configuration.workItems.increment();
-			}
-		}
-
-		@Override
-		public void doTask() throws EvaluationException {
-			long start = 0;
-			if (recordRuleDiagnostics) {
-				start = System.currentTimeMillis();
-			}
-			Iterable<Term[]> tups = it.next();
-			if (it.hasNext()) {
-				exec.recursivelyAddTask(new RuleSuffixEvaluator(rule, head, body, startPos, s.copy(), it));
-			}
-			try {
-				for (Term[] tup : tups) {
-					evaluate(tup);
-				}
-			} catch (UncheckedEvaluationException e) {
-				throw new EvaluationException(
-						"Exception raised while evaluating the rule: " + rule + "\n\n" + e.getMessage());
-			}
-			if (recordRuleDiagnostics) {
-				long end = System.currentTimeMillis();
-				Configuration.recordRuleSuffixTime(rule, end - start);
-			}
-		}
-
-		void evaluate(Term[] ans) throws UncheckedEvaluationException {
-			SimplePredicate p = (SimplePredicate) body[startPos];
-			updateBinding(p, ans);
-			int pos = startPos + 1;
-			@SuppressWarnings("unchecked")
-			Iterator<Term[]>[] stack = new Iterator[rule.getBodySize()];
-			boolean movingRight = true;
-			while (pos > startPos) {
-				if (pos == body.length) {
-					try {
-						reportFact(head.getSymbol(), head.getArgs(), s);
-					} catch (EvaluationException e) {
-						throw new UncheckedEvaluationException(
-								"Exception raised while evaluating the literal: " + head + "\n\n" + e.getMessage());
-					}
-					pos--;
-					movingRight = false;
-				} else if (movingRight) {
-					SimpleLiteral l = body[pos];
-					try {
-						switch (l.getTag()) {
-						case ASSIGNMENT:
-							((Assignment) l).assign(s);
-							pos++;
-							break;
-						case CHECK:
-							if (((Check) l).check(s)) {
-								pos++;
-							} else {
-								pos--;
-								movingRight = false;
-							}
-							break;
-						case DESTRUCTOR:
-							if (((Destructor) l).destruct(s)) {
-								pos++;
-							} else {
-								pos--;
-								movingRight = false;
-							}
-							break;
-						case PREDICATE:
-							Iterator<Iterable<Term[]>> tups = lookup(rule, pos, s).iterator();
-							if (((SimplePredicate) l).isNegated()) {
-								if (!tups.hasNext()) {
-									pos++;
-								} else {
-									pos--;
-									movingRight = false;
-								}
-							} else {
-								if (tups.hasNext()) {
-									stack[pos] = tups.next().iterator();
-									if (tups.hasNext()) {
-										exec.recursivelyAddTask(
-												new RuleSuffixEvaluator(rule, head, body, pos, s.copy(), tups));
-									}
-									// No need to do anything else: we'll hit the right case on the next iteration.
-								} else {
-									pos--;
-								}
-								movingRight = false;
-							}
-							break;
-						}
-					} catch (EvaluationException e) {
-						throw new UncheckedEvaluationException(
-								"Exception raised while evaluating the literal: " + l + "\n\n" + e.getMessage());
-					}
-				} else {
-					Iterator<Term[]> it = stack[pos];
-					if (it != null && it.hasNext()) {
-						ans = it.next();
-						updateBinding((SimplePredicate) rule.getBody(pos), ans);
-						movingRight = true;
-						pos++;
-					} else {
-						stack[pos] = null;
-						pos--;
-					}
-				}
-			}
-		}
-
-		void updateBinding(SimplePredicate p, Term[] ans) {
-			if (Configuration.recordWork) {
-				Configuration.work.increment();
-			}
-			Term[] args = p.getArgs();
-			BindingType[] pat = p.getBindingPattern();
-			for (int i = 0; i < pat.length; ++i) {
-				if (pat[i].isFree()) {
-					s.put((Var) args[i], ans[i]);
-				}
-			}
-		}
-
-	}
-
-	@SuppressWarnings("serial")
-	class RulePrefixEvaluator extends AbstractFJPTask {
-
-		final IndexedRule rule;
+		private final IndexedRule rule;
 
 		protected RulePrefixEvaluator(IndexedRule rule) {
 			super(exec);
@@ -387,7 +220,7 @@ public final class RoundBasedStratumEvaluator extends AbstractStratumEvaluator {
 			}
 		}
 
-		void evaluate() throws EvaluationException {
+		private void evaluate() throws EvaluationException {
 			int len = rule.getBodySize();
 			int pos = 0;
 			OverwriteSubstitution s = new OverwriteSubstitution();
@@ -442,7 +275,7 @@ public final class RoundBasedStratumEvaluator extends AbstractStratumEvaluator {
 		}
 	}
 
-	StopWatch recordRoundStart(int round) {
+	private StopWatch recordRoundStart(int round) {
 		if (!Configuration.debugRounds) {
 			return null;
 		}
@@ -455,7 +288,7 @@ public final class RoundBasedStratumEvaluator extends AbstractStratumEvaluator {
 		return watch;
 	}
 
-	void recordRoundEnd(int round, StopWatch watch) {
+	private void recordRoundEnd(int round, StopWatch watch) {
 		if (watch == null) {
 			return;
 		}
@@ -464,7 +297,7 @@ public final class RoundBasedStratumEvaluator extends AbstractStratumEvaluator {
 		System.err.println("Time: " + watch.getTime() + "ms");
 	}
 
-	StopWatch recordDbUpdateStart() {
+	private StopWatch recordDbUpdateStart() {
 		if (!Configuration.debugRounds) {
 			return null;
 		}
@@ -476,7 +309,7 @@ public final class RoundBasedStratumEvaluator extends AbstractStratumEvaluator {
 		return watch;
 	}
 
-	void recordDbUpdateEnd(StopWatch watch) {
+	private void recordDbUpdateEnd(StopWatch watch) {
 		if (watch == null) {
 			return;
 		}
